% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add function}
\usage{
add(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector which is a sum of x and y.
}
\description{
This function adds to numeric vector
}
\examples{
add (1, 1)
add (10, 1)
}
