% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnings.R
\docType{data}
\name{earnings}
\alias{earnings}
\title{Earnings of employees of XYZ company}
\format{
A data frame with 3 rows and 2 variables.
\describe{
 \item{name}{character - the name of an employee}
 \item{salary}{numerical - the monthly salary of the employee}
}
}
\usage{
data (earnings)
}
\description{
The data set contains the monthly salary of employees of company XYZ.
}
\examples{
barplot (earnings$salary, names.arg = earnings$name)
}
\keyword{datasets}
\keyword{earnings}
