% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_many_tables.R
\name{read_many_tables}
\alias{read_many_tables}
\title{Read many data frames with the same columns from many text files.}
\usage{
read_many_tables(prefix, range, suffix, is.header = TRUE)
}
\arguments{
\item{prefix}{A string with the last part of the filename. Typically consists of a file extension.}

\item{range}{Numeric vector with the numbers which differ from file to file.}

\item{is.header}{A local value indicating whether the file contains the names of the variables
as its first line. The default value is TRUE.}
}
\value{
A single data frame.
}
\description{
Read many data frames with the same columns from many text files.
}
\examples{
read.many.tables ("N1000_k8_O0.2_beta0.5_part", 1:10, ".txt")
read.many.tables ("precision_N", c(100,200,500,1000,2000,5000), ".txt")
}
